feat: move translation methods to global mixin
